Scrap delegation trait
